Replace useRef with useState for useSharedValue (#5458)
## Summary
The general idea here is that we can avoid the `makeMutable` call on
every single render. This change is also compatible with hot reloading.

We (at Discord) have been running this patch on top of our own
reanimated now for a few weeks and haven't noticed any side effects of
the change.

There has been a bit of additional discussion I've seen surrounding this
issue here:
https://github.com/software-mansion/react-native-reanimated/pull/3199

## Test plan
N/A