Improve reduced motion warning (#5487)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR extends the reduced motion warning to inform that it is only
displayed in development mode and adds a note in docs that explains why
the warning is displayed and how it can be disabled.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->