Improve TS in src [0/6] - Allow non-null assertion and remove some eslint directives (#5481)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Non-null assertions are considered to be a bad practice - however this
is in regard to pure TypeScript code. This is not the case in
Reanimated, since we have to deal with multiple platforms and multiple
threads. Therefore sometimes we know that a value is defined, without a
need for unnecessary checks for `undefined` or `null` - and non-null
assertions are very handy there, since they are very short.

Also, using `?` operator instead of TS `!` could lead to unexpected
behavior that could be hard to understand. Consider the following
example:

```TS
function dispatchCommandPaper<T extends Component>(
  animatedRef: AnimatedRef<T>,
  commandName: string,
  args: Array<unknown> = []
) {
  'worklet';
  if (!_WORKLET) {
    return;
  }

  const viewTag = animatedRef() as number;
  // TS error - _dispatchCommandPaper is not defined
  _dispatchCommandPaper(viewTag, commandName, args);
}
```

We get a TS error even though we are sure that `_dispatchCommandPaper`
is defined - `dispatchCommandPaper` is only used on Paper and
`_dispatchCommandPaper` is injected there. However, if we changed it to:

```TS
_dispatchCommandPaper?(viewTag, commandName, args);
```

And something would go so wrong that `_dispatchCommandPaper` would
actually be undefined this function would silently fail instead of
throwing with a meaningful stacktrace.

## Test plan

TS tests are sufficient for this.

Co-authored by: @marmor157.