Fix incomplete UI runtime initialization (#5505)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Fixes #5421, fixes #5331.

This PR removes `installValueUnpacker` method which was called on the JS
thread during the creation of `NativeReanimatedModule` but operated
directly on the UI runtime. Because of this, UI runtime was already used
from JS thread so any use from the UI runtime (e.g. Reanimated event
handlers) would cause a crash due to a failed check in
`ReanimatedReentrancyCheck::before`. Note that the check also occurred
in release mode due to improperly set `NDEBUG` flag.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->