Improve TS in src [5/6] - Acknowledge UI type of AnimatedRef (#5486)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently `useAnimatedRef` is an object that is of certain type on JS
thread but a significantly other on the UI thread. This UI thread type
is added and used in every relevant place.

## Test plan

CI

Co-authored by: @marmor157.