Allow layout animations to be updated on rerender (#5371)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR changes the current implementation to allow changes in layout
animations on rerender.

This PR is the first part of the layout animation configuration
overhaul. In a next PR I will extend those changes to SETs and remove
`configureLayoutAnimations` altogether.

TODO:

- [x] make it work
- [x] add example
- [x] add batchinator
- [x] take care of web
- [x] docs
- [x] fix issues with changing animations when an animation is already
running

# Issues

### Simultaneous layout animations

I encountered an interesting issue, that I was able to reproduce with
the following custom layout animation function on the main branch (so
without my changes).
```js
function CustomLayoutTransiton(): LayoutAnimationFunction {
  const isEven = makeMutable(1);
  return (values) => {
    'worklet';
    const isEvenLocal = isEven.value;
    isEven.value = 1 - isEven.value;

    if (isEvenLocal) {
      return {
        animations: {
          opacity: withTiming(1, { duration: 1000 }),
        },
        initialValues: {
          opacity: 0,
        },
      };
    }
    return {
      animations: {
        originX: values.targetOriginX,
      },
      initialValues: {
        originX: values.currentOriginX,
      },
    };
  };
}
```
The issue is that if a layout animation is started before the previous
animation finished, then the previous animation is cancelled. If those
animations operate on different sets of props, then we end up with
broken results. In this example we have the opacity of the box stay
lower than 1, if the second animation is requested too quickly.


https://github.com/software-mansion/react-native-reanimated/assets/56109050/ef18cea5-4a8b-44cd-b8c4-2fc5c6eb677e

In this PR I changed the implementation of `animationsManager` to merge
the layout animations. This is the same solution that was previously
implemented to avoid conflicts between entering and layout animations.

### Web

There are some issues with `visibility:hidden` on web, but they should
be fixed after #5278 gets merged

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->