fix: lazily search for ReactFabric (#5588)
## Summary

Quick PR fixing a problem that can be spotted in `Expensfiy` app where
for some reason on new arch, this code was evaluated before the
`nativeFabricUIManager` was added to global, making the code throw. Now
it is requested only when used for the first time, making it more safe.

## Test plan

For some reason I could not reproduce it in the `FabricExample` in any
way.