Introduce executeOnUIRuntimeSync and use it to replace Sync Data Holder (#4300)
## Summary

This PR introduces a new concept of executing worklets on the UI runtime
synchronously from the RN JS runtime.

It is implemented using runtime decorator that uses a lock to guard all
the code that's being executed on that runtime. Thanks to this locking
mechanism, we can now grab that lock on a different thread and safely
execute some more methods without worrying that the runtime object will
be accessed concurrently.

This new method makes opens up some new possibilities. The first one
being the way we handle two-way shared values. The shared value rewrite
introduced a concept of one-way shared values that are only accessed and
modified from the ui runtime. We initially thought such values can be
used in majority of the use cases, however, in practice it turned out
that a lot of existing code already relies on reading from shared values
on the RN thread. Despite the reads happening very infrequently (just a
few times per shared value runtime), we still needed to serialize all
the updates for a given shared value such that it can be read later on
from the RN thread. With this change, we can leverage synchronous
execution of the code on UI runtime and ask it to read the value on
demand rather than synchronizing the value after each single update.


## Test plan

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>
Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>